Return all PersonOversiktStatus with oppfolgingsplanLPSBistandUbehandlet=true
All persons with oppfolgingsplanLPSBistandUbehandlet=true should be returned in hentPersonoversiktStatusTilknyttetEnhet even though no Oppfolgingstilfelle exists on the PersonOversiktStatus. OppfolgingsplanLPSBistand that are Ubehandle should be shown to a Veileder of a Enhet because it is not a requirement for a person to be sick, to have been sick or to have received Sykepenger.