Close open motebehov that are closed in syfomotebehov
Set true motebehov to false on users with no motebehov in syfomotebehov.

We don't know why these motebehov_ubehandlet are true. Four of the users
don't even exist in syfomotebehov. Others haven't been sick in years.
Theories that have been checked out:

* Changed identities: PDL only had one id on the users we checked
* Saving in syfomotebehov failed, but the entry still ended up on Kafka:
not possible
* Scripts/code that wrongfully deleted entries i syfomotebehov: nothing
that have been committed

The list we're setting to false here is based on a query made across the
databases of the two apps, it contains all the current erronous true
ubehandlet_syfomotebehov. The plan now is to wait for new entries. It
will hopefully be easier to debug the error when it happens in a more
current state of the environment.